backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
mysqlclient
Flask-JWT-Extended
Flask-Marshmallow
Flask-CORS